dependency circles removed
