+++ css/train-theme.css
@@ -41,3 +41,16 @@
   color: rgb(218, 215, 215);
   font-size: 1.5em;
 }
+.play-column{
+  text-align: center;
+}
+.train-column {
+  background: #343a40;
+  color: #fff;
+  font-weight: bold;
+  text-align: center;
+}
+.train-column .dropbtn{
+  color: #fff;
+  font-weight: bold;
+}
